Module Css_parser

module Lex_buffer : sig ... end

A custom lexbuffer that automatically keeps track of the source location. This module is a thin wrapper arounds sedlexing's default buffer, which does not provide this functionality.

module Lexer : sig ... end

module Menhir_parser : sig ... end

module Parser : sig ... end

module Types : sig ... end