Module Css_parser.Lexer

val __sedlex_table_56 : string

val __sedlex_table_65 : string

val __sedlex_table_72 : string

val __sedlex_table_31 : string

val __sedlex_table_32 : string

val __sedlex_table_82 : string

val __sedlex_table_18 : string

val __sedlex_table_38 : string

val __sedlex_table_73 : string

val __sedlex_table_76 : string

val __sedlex_table_15 : string

val __sedlex_table_20 : string

val __sedlex_table_42 : string

val __sedlex_table_37 : string

val __sedlex_table_13 : string

val __sedlex_table_2 : string

val __sedlex_table_52 : string

val __sedlex_table_57 : string

val __sedlex_table_77 : string

val __sedlex_table_81 : string

val __sedlex_table_83 : string

val __sedlex_table_4 : string

val __sedlex_table_1 : string

val __sedlex_table_3 : string

val __sedlex_table_6 : string

val __sedlex_table_7 : string

val __sedlex_table_9 : string

val __sedlex_table_10 : string

val __sedlex_table_11 : string

val __sedlex_table_12 : string

val __sedlex_table_19 : string

val __sedlex_table_23 : string

val __sedlex_table_24 : string

val __sedlex_table_26 : string

val __sedlex_table_29 : string

val __sedlex_table_30 : string

val __sedlex_table_33 : string

val __sedlex_table_34 : string

val __sedlex_table_35 : string

val __sedlex_table_36 : string

val __sedlex_table_39 : string

val __sedlex_table_40 : string

val __sedlex_table_41 : string

val __sedlex_table_45 : string

val __sedlex_table_46 : string

val __sedlex_table_47 : string

val __sedlex_table_49 : string

val __sedlex_table_50 : string

val __sedlex_table_51 : string

val __sedlex_table_54 : string

val __sedlex_table_55 : string

val __sedlex_table_58 : string

val __sedlex_table_59 : string

val __sedlex_table_60 : string

val __sedlex_table_62 : string

val __sedlex_table_64 : string

val __sedlex_table_66 : string

val __sedlex_table_68 : string

val __sedlex_table_69 : string

val __sedlex_table_70 : string

val __sedlex_table_74 : string

val __sedlex_table_79 : string

val __sedlex_table_80 : string

val __sedlex_table_84 : string

val __sedlex_table_86 : string

val __sedlex_table_44 : string

val __sedlex_table_53 : string

val __sedlex_table_14 : string

val __sedlex_table_22 : string

val __sedlex_table_28 : string

val __sedlex_table_85 : string

val __sedlex_table_43 : string

val __sedlex_table_17 : string

val __sedlex_table_71 : string

val __sedlex_table_8 : string

val __sedlex_table_21 : string

val __sedlex_table_25 : string

val __sedlex_table_67 : string

val __sedlex_table_75 : string

val __sedlex_table_78 : string

val __sedlex_table_5 : string

val __sedlex_table_16 : string

val __sedlex_table_27 : string

val __sedlex_table_48 : string

val __sedlex_table_63 : string

val __sedlex_table_87 : string

val __sedlex_table_61 : string

val __sedlex_partition_82 : int -> int

val __sedlex_partition_35 : int -> int

val __sedlex_partition_56 : int -> int

val __sedlex_partition_31 : int -> int

val __sedlex_partition_25 : int -> int

val __sedlex_partition_61 : int -> int

val __sedlex_partition_91 : int -> int

val __sedlex_partition_16 : int -> int

val __sedlex_partition_44 : int -> int

val __sedlex_partition_90 : int -> int

val __sedlex_partition_78 : int -> int

val __sedlex_partition_45 : int -> int

val __sedlex_partition_67 : int -> int

val __sedlex_partition_30 : int -> int

val __sedlex_partition_94 : int -> int

val __sedlex_partition_96 : int -> int

val __sedlex_partition_11 : int -> int

val __sedlex_partition_15 : int -> int

val __sedlex_partition_29 : int -> int

val __sedlex_partition_38 : int -> int

val __sedlex_partition_41 : int -> int

val __sedlex_partition_43 : int -> int

val __sedlex_partition_95 : int -> int

val __sedlex_partition_66 : int -> int

val __sedlex_partition_42 : int -> int

val __sedlex_partition_52 : int -> int

val __sedlex_partition_26 : int -> int

val __sedlex_partition_51 : int -> int

val __sedlex_partition_17 : int -> int

val __sedlex_partition_64 : int -> int

val __sedlex_partition_65 : int -> int

val __sedlex_partition_54 : int -> int

val __sedlex_partition_7 : int -> int

val __sedlex_partition_37 : int -> int

val __sedlex_partition_8 : int -> int

val __sedlex_partition_19 : int -> int

val __sedlex_partition_83 : int -> int

val __sedlex_partition_58 : int -> int

val __sedlex_partition_60 : int -> int

val __sedlex_partition_1 : int -> int

val __sedlex_partition_36 : int -> int

val __sedlex_partition_85 : int -> int

val __sedlex_partition_57 : int -> int

val __sedlex_partition_62 : int -> int

val __sedlex_partition_2 : int -> int

val __sedlex_partition_6 : int -> int

val __sedlex_partition_13 : int -> int

val __sedlex_partition_32 : int -> int

val __sedlex_partition_59 : int -> int

val __sedlex_partition_46 : int -> int

val __sedlex_partition_80 : int -> int

val __sedlex_partition_18 : int -> int

val __sedlex_partition_88 : int -> int

val __sedlex_partition_50 : int -> int

val __sedlex_partition_55 : int -> int

val __sedlex_partition_73 : int -> int

val __sedlex_partition_34 : int -> int

val __sedlex_partition_24 : int -> int

val __sedlex_partition_49 : int -> int

val __sedlex_partition_93 : int -> int

val __sedlex_partition_69 : int -> int

val __sedlex_partition_70 : int -> int

val __sedlex_partition_14 : int -> int

val __sedlex_partition_48 : int -> int

val __sedlex_partition_84 : int -> int

val __sedlex_partition_74 : int -> int

val __sedlex_partition_71 : int -> int

val __sedlex_partition_81 : int -> int

val __sedlex_partition_27 : int -> int

val __sedlex_partition_53 : int -> int

val __sedlex_partition_63 : int -> int

val __sedlex_partition_79 : int -> int

val __sedlex_partition_9 : int -> int

val __sedlex_partition_20 : int -> int

val __sedlex_partition_75 : int -> int

val __sedlex_partition_76 : int -> int

val __sedlex_partition_89 : int -> int

val __sedlex_partition_4 : int -> int

val __sedlex_partition_33 : int -> int

val __sedlex_partition_10 : int -> int

val __sedlex_partition_23 : int -> int

val __sedlex_partition_86 : int -> int

val __sedlex_partition_98 : int -> int

val __sedlex_partition_21 : int -> int

val __sedlex_partition_39 : int -> int

val __sedlex_partition_97 : int -> int

val __sedlex_partition_92 : int -> int

val __sedlex_partition_3 : int -> int

val __sedlex_partition_47 : int -> int

val __sedlex_partition_40 : int -> int

val __sedlex_partition_12 : int -> int

val __sedlex_partition_22 : int -> int

val __sedlex_partition_72 : int -> int

val __sedlex_partition_5 : int -> int

val __sedlex_partition_87 : int -> int

val __sedlex_partition_68 : int -> int

val __sedlex_partition_77 : int -> int

val __sedlex_partition_28 : int -> int

CSS lexer. * Reference: * https://www.w3.org/TR/css-syntax-3/ * https://github.com/yahoo/css-js/blob/master/src/l/css.3.l

module Sedlexing = Lex_buffer

exception LexingError of Stdlib.Lexing.position * string

Signals a lexing error at the provided source location.

exception ParseError of Menhir_parser.token * Stdlib.Lexing.position * Stdlib.Lexing.position

Signals a parsing error at the provided token and its start and end * locations.

exception GrammarError of string * Location.t

Signals a grammar error at the provided location.

val position_to_string : Stdlib.Lexing.position -> string

val location_to_string : Location.t -> string

val dimension_to_string : Types.dimension -> string

val token_to_string : Menhir_parser.token -> string

val discard_comments_and_white_spaces : Sedlexing.t -> bool

val get_next_tokens : Sedlexing.t -> bool -> Menhir_parser.token list

val get_dimension : string -> Sedlexing.t -> Menhir_parser.token

val get_url : string -> Sedlexing.t -> Menhir_parser.token

val token_queue : (Menhir_parser.token * Stdlib.Lexing.position * Stdlib.Lexing.position) Stdlib.Queue.t

val queue_next_tokens_with_location : Sedlexing.t -> unit

val parse : Sedlexing.t -> ( Menhir_parser.token, 'a ) MenhirLib.Convert.traditional -> 'a

val parse_string : ?container_lnum:int -> ?pos:Stdlib.Lexing.position -> string -> ( Menhir_parser.token, 'a ) MenhirLib.Convert.traditional -> 'a